feat: second kill products
